Move the Index...Bad funcs into a _test.go file in the keeper package so that they're only exposed for other unit tests.
